Move auth module to API package
